fix: use actual Radar RPC signature

+++ supabase/migrations/20260921235000_lock_down_radar_rpc.sql
@@ -1,3 +1,3 @@
 -- Radar is an authenticated, location-sensitive feature. Do not expose its RPC to anonymous callers.
-revoke execute on function public.nearby_radar(double precision, double precision, double precision, text) from anon;
+revoke execute on function public.nearby_radar(double precision, double precision, integer, text) from anon;
 grant execute on function public.nearby_radar(double precision, double precision, double precision, text) to authenticated;
